Move NetworkConnectivityChecker into hi_tools module
Christoph's preference: no separate module, integrate directly into hi_tools.

Changes:
- Add NetworkConnectivityChecker.h/.cpp and three platform impl files to
  hi_tools/hi_tools/, included from hi_tools_01.cpp
- Rewrite the Apple impl (NetworkConnectivityCheckerApple.cpp) using the
  SystemConfiguration C API directly -- eliminates the Objective-C
  Reachability wrapper and the need for any .mm file in hi_tools
- Add SystemConfiguration framework to hi_tools.h OSX/iOS framework lists
- Remove hi_network_connectivity from hi_scripting.h dependencies and from
  all .jucer and template files
- Delete the hi_network_connectivity module entirely
