+++ hi_backend/backend/ProjectTemplate.cpp
@@ -106,7 +106,6 @@ static const unsigned char projectTemplate_jucer_lines[] = R"(
         <MODULEPATH id="hi_streaming" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_tools" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_snex" path="%HISE_PATH%"/>
-        <MODULEPATH id="hi_network_connectivity" path="%HISE_PATH%"/>
 		<MODULEPATH id="melatonin_blur" path="%HISE_PATH%"/>
         %LORIS_MODULEPATH%
       </MODULEPATHS>
@@ -154,7 +153,6 @@ static const unsigned char projectTemplate_jucer_lines[] = R"(
         <MODULEPATH id="hi_streaming" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_tools" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_snex" path="%HISE_PATH%"/>
-        <MODULEPATH id="hi_network_connectivity" path="%HISE_PATH%"/>
 		<MODULEPATH id="melatonin_blur" path="%HISE_PATH%"/>
         %LORIS_MODULEPATH%
       </MODULEPATHS>
@@ -208,7 +206,6 @@ static const unsigned char projectTemplate_jucer_lines[] = R"(
         <MODULEPATH id="hi_streaming" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_tools" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_snex" path="%HISE_PATH%"/>
-        <MODULEPATH id="hi_network_connectivity" path="%HISE_PATH%"/>
         <MODULEPATH id="melatonin_blur" path="%HISE_PATH%"/>
         %LORIS_MODULEPATH%
       </MODULEPATHS>
@@ -257,7 +254,6 @@ static const unsigned char projectTemplate_jucer_lines[] = R"(
         <MODULEPATH id="hi_streaming" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_tools" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_snex" path="%HISE_PATH%"/>
-        <MODULEPATH id="hi_network_connectivity" path="%HISE_PATH%"/>
 		<MODULEPATH id="melatonin_blur" path="%HISE_PATH%"/>
         %LORIS_MODULEPATH%
       </MODULEPATHS>
@@ -301,7 +297,6 @@ static const unsigned char projectTemplate_jucer_lines[] = R"(
         <MODULEPATH id="hi_core" path="%HISE_PATH%"/>
         <MODULEPATH id="hi_tools" path="%HISE_PATH%"/>
         <MODULEPATH id="hnode_jit" path="%HISE_PATH%"/>
-        <MODULEPATH id="hi_network_connectivity" path="%HISE_PATH%"/>
 		<MODULEPATH id="melatonin_blur" path="%HISE_PATH%"/>
         %LORIS_MODULEPATH%
       </MODULEPATHS>
@@ -310,7 +305,6 @@ static const unsigned char projectTemplate_jucer_lines[] = R"(
   <MODULES>
     <MODULE id="hi_core" showAllCode="1" useLocalCopy="0"/>
     <MODULE id="hi_frontend" showAllCode="1" useLocalCopy="0"/>
-    <MODULE id="hi_network_connectivity" showAllCode="1" useLocalCopy="0"/>
     <MODULE id="hi_scripting" showAllCode="1" useLocalCopy="0"/>
     <MODULE id="hi_dsp_library" showAllCode="1" useLocalCopy="0"/>
     <MODULE id="hi_faust" showAllCode="1" useLocalCopy="0"/>
