--- hi_network_connectivity/hi_network_connectivity.h
@@ -1,59 +1,0 @@
-/*  ===========================================================================
-*
-*   This file is part of HISE.
-*   Copyright 2016 Christoph Hart
-*
-*   HISE is free software: you can redistribute it and/or modify
-*   it under the terms of the GNU General Public License as published by
-*   the Free Software Foundation, either version 3 of the License, or
-*   (at your option) any later version.
-*
-*   HISE is distributed in the hope that it will be useful,
-*   but WITHOUT ANY WARRANTY; without even the implied warranty of
-*   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-*   GNU General Public License for more details.
-*
-*   You should have received a copy of the GNU General Public License
-*   along with HISE.  If not, see <http://www.gnu.org/licenses/>.
-*
-*   Commercial licenses for using HISE in an closed source project are
-*   available on request. Please visit the project's website to get more
-*   information about commercial licensing:
-*
-*   http://www.hise.audio/
-*
-*   HISE is based on the JUCE library,
-*   which must be separately licensed for closed source applications:
-*
-*   http://www.juce.com
-*
-*   ===========================================================================
-*/
-
-/******************************************************************************
-
-BEGIN_JUCE_MODULE_DECLARATION
-
-  ID:               hi_network_connectivity
-  vendor:           Hart Instruments
-  version:          1.0.0
-  name:             HISE Network Connectivity
-  description:      Native OS network connectivity monitoring without polling external servers
-  website:          http://hise.audio
-  license:          GPL / Commercial
-
-  dependencies:     juce_core, juce_events
-
-  OSXFrameworks:    SystemConfiguration
-  iOSFrameworks:    SystemConfiguration
-
-END_JUCE_MODULE_DECLARATION
-
-******************************************************************************/
-
-#pragma once
-
-#include <juce_core/juce_core.h>
-#include <juce_events/juce_events.h>
-
-#include "NetworkConnectivityChecker.h"
