--- hi_network_connectivity/hi_network_connectivity.mm
@@ -1,1 +1,0 @@
-#include "hi_network_connectivity.cpp"
